Move all keys to key.hpp
